+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added methods to ListExtensions.
+
+### Changed
+- Changed Rescale to RescaleLength.
 
 ## [1.5.0] - 2019-10-11
 ### Added